+++ public/index.html
@@ -143,6 +143,6 @@
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
